requirements.txt
configparser==3.5.0
pycrypto==2.6.1
scapy==2.3.3
